requirements.txt
neptune==1.2.0
neptune-pytorch==0.2.0
matplotlib==3.7.1
numpy==1.21.2
pandas==1.5.0
pillow==9.3.0
scipy==1.9.1
configparser==7.0.0
torchsummary==1.5.1
tqdm==4.64.1
ultralytics
pycocotools==2.0.8